+++ FluentFTP/Client/FtpClient_FolderUpload.cs
@@ -93,7 +93,7 @@ namespace FluentFTP {
 			UploadDirectoryFiles(filesToUpload, existsMode, verifyOptions, progress, remoteListing);
 
 			// delete the extra remote files if in mirror mode and the directory was pre-existing
-			DeleteExtraServerFiles(mode, shouldExist, remoteListing);
+			DeleteExtraServerFiles(mode, remoteFolder, shouldExist, remoteListing, rules);
 
 			return results;
 		}
@@ -199,7 +199,7 @@ namespace FluentFTP {
 			await UploadDirectoryFilesAsync(filesToUpload, existsMode, verifyOptions, progress, remoteListing, token);
 
 			// delete the extra remote files if in mirror mode and the directory was pre-existing
-			await DeleteExtraServerFilesAsync(mode, shouldExist, remoteListing, token);
+			await DeleteExtraServerFilesAsync(mode, remoteFolder, shouldExist, remoteListing, rules, token);
 
 			return results;
 		}
@@ -235,7 +235,7 @@ namespace FluentFTP {
 				if (!FilePassesRules(result, rules, true)) {
 					continue;
 				}
-				
+
 				dirsToUpload.Add(result);
 			}
 
@@ -329,16 +329,9 @@ namespace FluentFTP {
 
 				// skip uploading the file if it does not pass all the rules
 				if (!FilePassesRules(result, rules, true)) {
-
-					// record that this file/folder should exist
-					// if we don't want to delete excluded files
-					if (!UploadDirectoryDeleteExcluded) {
-						shouldExist.Add(remoteFile.ToLower(), true);
-					}
-
 					continue;
 				}
-				
+
 				// record that this file should exist
 				shouldExist.Add(remoteFile.ToLower(), true);
 
@@ -365,7 +358,7 @@ namespace FluentFTP {
 
 					// skip uploading if the file already exists on the server
 					FtpRemoteExists existsModeToUse;
-					if (!CanUploadFile(result, remoteListing, existsMode, out existsModeToUse)){
+					if (!CanUploadFile(result, remoteListing, existsMode, out existsModeToUse)) {
 						continue;
 					}
 
@@ -462,7 +455,7 @@ namespace FluentFTP {
 		/// <summary>
 		/// Delete the extra remote files if in mirror mode and the directory was pre-existing
 		/// </summary>
-		private void DeleteExtraServerFiles(FtpFolderSyncMode mode, Dictionary<string, bool> shouldExist, FtpListItem[] remoteListing) {
+		private void DeleteExtraServerFiles(FtpFolderSyncMode mode, string remoteFolder, Dictionary<string, bool> shouldExist, FtpListItem[] remoteListing, List<FtpRule> rules) {
 			if (mode == FtpFolderSyncMode.Mirror && remoteListing != null) {
 
 				LogFunc(nameof(DeleteExtraServerFiles));
@@ -474,14 +467,16 @@ namespace FluentFTP {
 
 						if (!shouldExist.ContainsKey(existingServerFile.FullName.ToLower())) {
 
-							LogStatus(FtpTraceLevel.Info, "Delete extra file from server: " + existingServerFile.FullName);
+							// only delete the remote file if its permitted by the configuration
+							if (CanDeleteRemoteFile(rules, existingServerFile)) {
+								LogStatus(FtpTraceLevel.Info, "Delete extra file from server: " + existingServerFile.FullName);
 
-							// delete the file from the server
-							try {
-								DeleteFile(existingServerFile.FullName);
+								// delete the file from the server
+								try {
+									DeleteFile(existingServerFile.FullName);
+								}
+								catch (Exception ex) { }
 							}
-							catch (Exception ex) { }
-
 						}
 
 					}
@@ -495,7 +490,7 @@ namespace FluentFTP {
 		/// <summary>
 		/// Delete the extra remote files if in mirror mode and the directory was pre-existing
 		/// </summary>
-		private async Task DeleteExtraServerFilesAsync(FtpFolderSyncMode mode, Dictionary<string, bool> shouldExist, FtpListItem[] remoteListing, CancellationToken token) {
+		private async Task DeleteExtraServerFilesAsync(FtpFolderSyncMode mode, string remoteFolder, Dictionary<string, bool> shouldExist, FtpListItem[] remoteListing, List<FtpRule> rules, CancellationToken token) {
 			if (mode == FtpFolderSyncMode.Mirror && remoteListing != null) {
 
 				LogFunc(nameof(DeleteExtraServerFilesAsync));
@@ -507,14 +502,16 @@ namespace FluentFTP {
 
 						if (!shouldExist.ContainsKey(existingServerFile.FullName.ToLower())) {
 
-							LogStatus(FtpTraceLevel.Info, "Delete extra file from server: " + existingServerFile.FullName);
+							// only delete the remote file if its permitted by the configuration
+							if (CanDeleteRemoteFile(rules, existingServerFile)) {
+								LogStatus(FtpTraceLevel.Info, "Delete extra file from server: " + existingServerFile.FullName);
 
-							// delete the file from the server
-							try {
-								await DeleteFileAsync(existingServerFile.FullName, token);
+								// delete the file from the server
+								try {
+									await DeleteFileAsync(existingServerFile.FullName, token);
+								}
+								catch (Exception ex) { }
 							}
-							catch (Exception ex) { }
-
 						}
 
 					}
@@ -523,8 +520,41 @@ namespace FluentFTP {
 
 			}
 		}
+
 #endif
 
+		/// <summary>
+		/// Check if the remote file can be deleted, based on the UploadDirectoryDeleteExcluded property
+		/// </summary>
+		private bool CanDeleteRemoteFile(List<FtpRule> rules, FtpListItem existingServerFile) {
+
+			// if we should not delete excluded files
+			if (!UploadDirectoryDeleteExcluded && !rules.IsBlank()) {
+
+				// create the result object to validate rules to ensure that file from excluded
+				// directories are not deleted on the FTP remote server
+				var result = new FtpResult() {
+					Type = existingServerFile.Type,
+					Size = existingServerFile.Size,
+					Name = Path.GetFileName(existingServerFile.FullName),
+					RemotePath = existingServerFile.FullName,
+					IsDownload = false,
+				};
+
+				// check if the file passes the rules
+				if (FilePassesRules(result, rules, false)) {
+					// delete the file because it is included
+					return true;
+				}
+				else {
+					// do not delete the file because it is excluded
+					return false;
+				}
+			}
+
+			// always delete the file whether its included or excluded by the rules
+			return true;
+		}
 
 	}
 }
